[Refactor] query수정, 불필요한 주석 삭제
