+++ src/main/java/me/snaptime/user/data/controller/UserController.java
@@ -10,8 +10,8 @@ import me.snaptime.user.data.dto.request.UserReqDto;
 import me.snaptime.user.data.dto.request.UserUpdateDto;
 import me.snaptime.user.data.dto.response.SignInResDto;
 import me.snaptime.user.data.dto.response.UserResDto;
-import me.snaptime.user.data.dto.response.userprofile2.AlbumSnapResDto;
-import me.snaptime.user.data.dto.response.userprofile2.UserProfileResDto;
+import me.snaptime.user.data.dto.response.userprofile.AlbumSnapResDto;
+import me.snaptime.user.data.dto.response.userprofile.UserProfileResDto;
 import me.snaptime.user.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
