+++ src/main/java/me/snaptime/user/data/dto/response/userprofile/UserProfileResDto.java
@@ -1,4 +1,4 @@
-package me.snaptime.user.data.dto.response.userprofile2;
+package me.snaptime.user.data.dto.response.userprofile;
 
 import lombok.Builder;
 import me.snaptime.user.data.domain.User;
