+++ src/main/java/me/snaptime/user/data/repository/UserCustomRepository.java
@@ -1,13 +1,11 @@
 package me.snaptime.user.data.repository;
 
 import me.snaptime.user.data.domain.User;
-import me.snaptime.user.data.dto.response.userprofile2.AlbumSnapResDto;
+import me.snaptime.user.data.dto.response.userprofile.AlbumSnapResDto;
 
 import java.util.List;
 
 public interface UserCustomRepository {
-    //UserProfileResDto findUserProfile1(User reqUser);
-    //UserProfileResDto findUserProfile2(User reqUser);
     List<AlbumSnapResDto> fidAlbumSnap(User reqUser);
 
 }
