+++ src/main/java/me/snaptime/user/service/UserService.java
@@ -13,8 +13,8 @@ import me.snaptime.user.data.dto.request.UserReqDto;
 import me.snaptime.user.data.dto.request.UserUpdateDto;
 import me.snaptime.user.data.dto.response.SignInResDto;
 import me.snaptime.user.data.dto.response.UserResDto;
-import me.snaptime.user.data.dto.response.userprofile2.AlbumSnapResDto;
-import me.snaptime.user.data.dto.response.userprofile2.UserProfileResDto;
+import me.snaptime.user.data.dto.response.userprofile.AlbumSnapResDto;
+import me.snaptime.user.data.dto.response.userprofile.UserProfileResDto;
 import me.snaptime.user.data.repository.ProfilePhotoRepository;
 import me.snaptime.user.data.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Value;
